+++ CHANGELOG.md
@@ -8,7 +8,7 @@ All notable changes to this project will be documented in this file.
 - New {idnumber} tag.
 - New {fa...} tag (for FontAwesome).
 - New {glyphicon...} tag (for Glyphicons).
-- New {sessionid} and %7Bsessionid%7D tags.
+- New {sectionid} and %7Bsectionid%7D tags.
 - New {details}, {summary}, {/summary}, {/details} tags.
 ### Updated
 - Fixed {categories} filter code compatibility with Moodle 2.7 to 3.5.